Move package of DistSQL exception
